requirements.txt
python-telegram-bot>=21,<23
openai>=1,<3
python-dotenv>=1,<2
APScheduler>=3,<4
tzdata>=2025,<2027
